% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars_manipulation.R
\name{average_mpg}
\alias{average_mpg}
\title{average_mpg}
\usage{
average_mpg(tbl)
}
\arguments{
\item{tbl}{}
}
\description{
This function is an example of pure function, that requires a package to
operates: dplyr in this case
}
