% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division.R
\name{my_division}
\alias{my_division}
\title{my_division}
\usage{
my_division(x, y)
}
\arguments{
\item{x}{\link{numeric} numerator}

\item{y}{\link{numeric} denominator}
}
\value{
\link{numeric} result of the division
}
\description{
This function is an example of pure function. it is exported: the user of the
package will be able to use it. It does not need to import packages: it does
not require any other packages to operates.
}
