% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division.R
\name{my_division}
\alias{my_division}
\title{my_division}
\usage{
my_division(x, y)
}
\arguments{
\item{x}{numeric numerator}

\item{y}{numeric denominator}
}
\value{
numeric result of the division
}
\description{
This function divides x by y
}
